requirements.txt
aiogram==3.0.0b7
telethon==1.30.0
pymongo==4.6.1
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
cryptography==41.0.3
